requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bertopic==0.3.1
bleach==6.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
docopt==0.6.2
filelock==3.13.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.12.2
hdbscan==0.8.33
huggingface-hub==0.20.2
idna==3.4
Jinja2==3.1.3
joblib==1.3.2
jsonable==0.3.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
llvmlite==0.41.1
lxml==5.1.0
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mwapi==0.6.1
mwcli==0.0.3
mwtypes==0.3.2
mwviews==0.2.1
mwxml==0.3.3
networkx==3.2
nltk==3.8.1
numba==0.58.1
numpy==1.26.0
packaging==23.2
pandas==2.1.1
para==0.0.8
patsy==0.5.3
Pillow==10.1.0
plotly==5.19.0
pynndescent==0.5.11
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rpds-py==0.10.6
safetensors==0.4.1
scikit-learn==1.4.1.post1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
soupsieve==2.5
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.3
umap-learn==0.5.5
urllib3==2.0.5
webencodings==0.5.1
wikipedia==1.4.0
yarl==1.9.2
